[CHORE]#61 : EditViewModel -> EntireEditViewModel 변경
