+++ feature/entire/src/main/java/com/bff/wespot/entire/screen/viewmodel/EntireViewModel.kt
@@ -2,31 +2,14 @@ package com.bff.wespot.entire.screen.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.bff.wespot.designsystem.component.indicator.WSToastType
-import com.bff.wespot.domain.repository.CommonRepository
 import com.bff.wespot.domain.repository.auth.AuthRepository
 import com.bff.wespot.domain.repository.message.MessageRepository
 import com.bff.wespot.domain.repository.message.MessageStorageRepository
 import com.bff.wespot.domain.repository.user.UserRepository
-import com.bff.wespot.domain.usecase.CheckProfanityUseCase
-import com.bff.wespot.entire.R
-import com.bff.wespot.entire.screen.common.INPUT_DEBOUNCE_TIME
-import com.bff.wespot.entire.screen.common.INTRODUCTION_MAX_LENGTH
 import com.bff.wespot.entire.screen.state.EntireAction
 import com.bff.wespot.entire.screen.state.EntireSideEffect
 import com.bff.wespot.entire.screen.state.EntireUiState
-import com.bff.wespot.model.ToastState
-import com.bff.wespot.model.common.BackgroundColor
-import com.bff.wespot.model.common.Character
-import com.bff.wespot.model.user.response.ProfileCharacter
 import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.debounce
-import kotlinx.coroutines.flow.distinctUntilChanged
-import kotlinx.coroutines.flow.flow
-import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
 import org.orbitmvi.orbit.ContainerHost
 import org.orbitmvi.orbit.syntax.simple.intent
@@ -42,54 +25,19 @@ class EntireViewModel @Inject constructor(
     private val authRepository: AuthRepository,
     private val messageRepository: MessageRepository,
     private val messageStorageRepository: MessageStorageRepository,
-    private val commonRepository: CommonRepository,
-    private val checkProfanityUseCase: CheckProfanityUseCase,
 ) : ViewModel(), ContainerHost<EntireUiState, EntireSideEffect> {
     override val container = container<EntireUiState, EntireSideEffect>(EntireUiState())
 
-    private val introductionInput: MutableStateFlow<String> = MutableStateFlow("")
-
-    val characters: StateFlow<List<Character>> = flow {
-        commonRepository.getCharacters().onSuccess { emit(it) }
-    }.stateIn(
-        scope = viewModelScope,
-        started = SharingStarted.WhileSubscribed(5000),
-        initialValue = emptyList(),
-    )
-
-    val backgroundColor: StateFlow<List<BackgroundColor>> = flow {
-        commonRepository.getBackgroundColors().onSuccess { emit(it) }
-    }.stateIn(
-        scope = viewModelScope,
-        started = SharingStarted.WhileSubscribed(5000),
-        initialValue = emptyList(),
-    )
-
     fun onAction(action: EntireAction) {
         when (action) {
-            EntireAction.OnEntireScreenEntered,
-            EntireAction.OnRevokeScreenEntered,
-            EntireAction.OnCharacterEditScreenEntered,
-            -> getProfile()
+            EntireAction.OnEntireScreenEntered, EntireAction.OnRevokeScreenEntered -> getProfile()
             EntireAction.OnRevokeButtonClicked -> revokeUser()
             EntireAction.OnSignOutButtonClicked -> signOut()
             EntireAction.OnRevokeConfirmed -> handleRevokeConfirmed()
-            EntireAction.OnIntroductionEditDoneButtonClicked -> updateIntroduction()
             EntireAction.OnBlockListScreenEntered -> getUnBlockedMessage()
             EntireAction.UnBlockMessage -> unblockMessage()
-            is EntireAction.OnProfileEditScreenEntered -> {
-                getProfile()
-                observeIntroductionInput()
-                if (action.isCompleteEdit) {
-                    postEditDoneSideToast()
-                }
-            }
-            is EntireAction.OnProfileEditTextFieldFocused ->
-                handleProfileEditButtonText(action.focused)
             is EntireAction.OnUnBlockButtonClicked -> handleUnBlockButtonClicked(action.messageId)
             is EntireAction.OnRevokeReasonSelected -> handleRevokeReasonSelected(action.reason)
-            is EntireAction.OnIntroductionChanged -> handleIntroductionChanged(action.introduction)
-            is EntireAction.OnCharacterEditDoneButtonClicked -> updateCharacter(action.character)
         }
     }
 
@@ -98,7 +46,6 @@ class EntireViewModel @Inject constructor(
             userRepository.getProfile()
                 .onSuccess { profile ->
                     reduce { state.copy(profile = profile) }
-                    handleIntroductionChanged(profile.introduction)
                 }
                 .onFailure {
                     Timber.e(it)
@@ -175,88 +122,9 @@ class EntireViewModel @Inject constructor(
         }
     }
 
-    private fun handleProfileEditButtonText(focused: Boolean) = intent {
-        reduce {
-            state.copy(isIntroductionEditing = focused)
-        }
-    }
-
     private fun handleRevokeConfirmed() = intent {
         reduce {
             state.copy(revokeConfirmed = state.revokeConfirmed.not())
-        }
-    }
-
-    private fun handleIntroductionChanged(introduction: String) = intent {
-        reduce {
-            introductionInput.value = introduction
-            state.copy(introductionInput = introduction)
-        }
-    }
-
-    private fun observeIntroductionInput() {
-        viewModelScope.launch {
-            introductionInput
-                .debounce(INPUT_DEBOUNCE_TIME)
-                .distinctUntilChanged()
-                .collect { introduction ->
-                    if (introduction.length in 1..INTRODUCTION_MAX_LENGTH) {
-                        hasProfanity(introduction)
-                    }
-                }
-        }
-    }
-
-    private fun postEditDoneSideToast() = intent {
-        postSideEffect(
-            EntireSideEffect.ShowToast(
-                ToastState(
-                    show = true,
-                    message = R.string.edit_done,
-                    type = WSToastType.Success,
-                ),
-            ),
-        )
-    }
-
-    private fun hasProfanity(introduction: String) = intent {
-        runCatching {
-            val result = checkProfanityUseCase(introduction)
-            reduce {
-                state.copy(
-                    hasProfanity = result,
-                )
-            }
-        }
-    }
-
-    private fun updateIntroduction() = intent {
-        reduce { state.copy(isLoading = true) }
-        viewModelScope.launch {
-            userRepository.updateIntroduction(state.introductionInput)
-                .onSuccess {
-                    postEditDoneSideToast()
-                    reduce { state.copy(isLoading = false) }
-                }
-                .onFailure {
-                    Timber.e(it)
-                    reduce { state.copy(isLoading = false) }
-                }
-        }
-    }
-
-    private fun updateCharacter(character: ProfileCharacter) = intent {
-        reduce { state.copy(isLoading = true) }
-        viewModelScope.launch {
-            userRepository.updateCharacter(character = character)
-                .onSuccess {
-                    postSideEffect(EntireSideEffect.NavigateToEntire)
-                    reduce { state.copy(isLoading = false) }
-                }
-                .onFailure {
-                    Timber.e(it)
-                    reduce { state.copy(isLoading = false) }
-                }
         }
     }
 }
